+++ crates/rust-client/src/store/web_store/js/sync.js
@@ -95,7 +95,7 @@ export async function applyStateSync(
     partialBlockchainPeaksAsFlattenedVec
   );
 
-  return db.transaction(
+  return await db.transaction(
     "rw",
     stateSync,
     inputNotes,
@@ -112,7 +112,7 @@ export async function applyStateSync(
           newBlockNums[i],
           newBlockHeaders[i],
           partialBlockchainPeaks[i],
-          hasClientNotes[i]
+          hasClientNotes[i] == 1 // hasClientNotes is a u8 array, so we convert it to boolean
         );
       }
       await updatePartialBlockchainNodes(tx, nodeIndexes, nodes);