+++ crates/rust-client/src/test_utils/common.rs
@@ -274,11 +274,12 @@ pub async fn wait_for_tx(client: &mut TestClient, transaction_id: TransactionId)
             .unwrap();
 
         match tracked_transaction.status {
-            TransactionStatus::Committed(_) => {
+            TransactionStatus::Committed(n) => {
+                println!("tx committed in {n}");
                 break;
             },
             TransactionStatus::Pending => {
-                std::thread::sleep(Duration::from_millis(100));
+                std::thread::sleep(Duration::from_secs(1));
             },
             TransactionStatus::Discarded(cause) => {
                 panic!("Transaction was discarded with cause: {:?}", cause);
@@ -318,7 +319,7 @@ pub async fn wait_for_blocks(client: &mut TestClient, amount_of_blocks: u32) -> 
             return summary;
         }
 
-        std::thread::sleep(Duration::from_millis(100));
+        std::thread::sleep(Duration::from_secs(3));
     }
 }