+++ crates/web-client/test/new_transactions.test.ts
@@ -13,6 +13,120 @@ import { Account, TransactionRecord } from "../dist/crates/miden_client_web";
 // NEW_MINT_TRANSACTION TESTS
 // =======================================================================================================
 
+interface MultipleMintsTransactionResult {
+  transactionIds: string[];
+  createdNoteIds: string[];
+  numOutputNotesCreated: number;
+  nonce: string | undefined;
+  finalBalance: string | undefined;
+}
+
+const multipleMintsTest = async (
+  targetAccount: string,
+  faucetAccount: string,
+  withRemoteProver: boolean = false
+): Promise<MultipleMintsTransactionResult> => {
+  return await testingPage.evaluate(
+    async (
+      _targetAccount: string,
+      _faucetAccount: string,
+      _withRemoteProver: boolean
+    ) => {
+      const client = window.client;
+
+      const targetAccountId = window.AccountId.fromHex(_targetAccount);
+      const faucetAccountId = window.AccountId.fromHex(_faucetAccount);
+      await client.syncState();
+
+      // Mint 3 notes
+      let result: {
+        transactionIds: string[];
+        createdNoteIds: string[];
+        numOutputNotesCreated: number;
+      } = {
+        transactionIds: [],
+        createdNoteIds: [],
+        numOutputNotesCreated: 0,
+      };
+
+      for (let i = 0; i < 3; i++) {
+        const mintTransactionRequest = await client.newMintTransactionRequest(
+          targetAccountId,
+          faucetAccountId,
+          window.NoteType.Public,
+          BigInt(1000)
+        );
+
+        const mintTransactionResult = await client.newTransaction(
+          faucetAccountId,
+          mintTransactionRequest
+        );
+
+        if (_withRemoteProver && window.remoteProverUrl != null) {
+          await client.submitTransaction(
+            mintTransactionResult,
+            window.remoteProverInstance
+          );
+        } else {
+          await client.submitTransaction(mintTransactionResult);
+        }
+
+        await window.helpers.waitForTransaction(
+          mintTransactionResult.executedTransaction().id().toHex()
+        );
+
+        result.createdNoteIds.push(
+          mintTransactionResult.createdNotes().notes()[0].id().toString()
+        );
+        result.transactionIds.push(
+          mintTransactionResult.executedTransaction().id().toHex()
+        );
+        result.numOutputNotesCreated += mintTransactionResult
+          .createdNotes()
+          .numNotes();
+      }
+
+      // Consume the minted notes
+      for (let i = 0; i < result.createdNoteIds.length; i++) {
+        const consumeTransactionRequest = client.newConsumeTransactionRequest([
+          result.createdNoteIds[i],
+        ]);
+        const consumeTransactionResult = await client.newTransaction(
+          targetAccountId,
+          consumeTransactionRequest
+        );
+
+        if (_withRemoteProver && window.remoteProverUrl != null) {
+          await client.submitTransaction(
+            consumeTransactionResult,
+            window.remoteProverInstance
+          );
+        } else {
+          await client.submitTransaction(consumeTransactionResult);
+        }
+
+        await window.helpers.waitForTransaction(
+          consumeTransactionResult.executedTransaction().id().toHex()
+        );
+      }
+
+      const changedTargetAccount = await client.getAccount(targetAccountId);
+
+      return {
+        ...result,
+        nonce: changedTargetAccount.nonce()?.toString(),
+        finalBalance: changedTargetAccount
+          .vault()
+          .getBalance(faucetAccountId)
+          .toString(),
+      };
+    },
+    targetAccount,
+    faucetAccount,
+    withRemoteProver
+  );
+};
+
 describe("mint transaction tests", () => {
   const testCases = [
     { flag: false, description: "mint transaction completes successfully" },
@@ -24,12 +138,18 @@ describe("mint transaction tests", () => {
 
   testCases.forEach(({ flag, description }) => {
     it(description, async () => {
+      // This test was added in #995 to reproduce an issue in the web wallet.
+      // It is useful because most tests consume the note right on the latest client block,
+      // but this test mints 3 notes and consumes them after the fact. This ensures the
+      // MMR data for old blocks is available and valid so that the notes can be consumed.
       const { faucetId, accountId } = await setupWalletAndFaucet();
-      const result = await mintTransaction(accountId, faucetId, flag);
+      const result = await multipleMintsTest(accountId, faucetId, flag);
 
-      expect(result.transactionId).to.not.be.empty;
-      expect(result.numOutputNotesCreated).to.equal(1);
-      expect(result.nonce).to.equal("1");
+      expect(result.transactionIds.length).to.equal(3);
+      expect(result.numOutputNotesCreated).to.equal(3);
+      expect(result.nonce).to.equal("3");
+      expect(result.finalBalance).to.equal("3000");
+      expect(result.createdNoteIds.length).to.equal(3);
     });
   });
 });